fix: isolate config tests from Kent persistence
